Fixed handling of bokeh not installed
